Fix infinite loop when Force Shield has 256 HP
(cherry picked from commit 128616d1aa18e529f1ba67c4fde05bc56266d8a9)

+++ src/st_stuff.c
@@ -1468,7 +1468,7 @@ static void ST_drawPowerupHUD(void)
 
 		if ((stplyr->powers[pw_shield] & SH_NOSTACK & ~SH_FORCEHP) == SH_FORCE)
 		{
-			UINT8 i, max = (stplyr->powers[pw_shield] & SH_FORCEHP);
+			UINT16 i, max = (stplyr->powers[pw_shield] & SH_FORCEHP);
 			for (i = 0; i <= max; i++)
 			{
 				V_DrawSmallScaledPatch(offs-(i<<1), hudinfo[HUD_POWERUPS].y-(i<<1), (V_PERPLAYER|hudinfo[HUD_POWERUPS].f)|((i == max) ? V_HUDTRANS : V_HUDTRANSHALF), forceshield);
